requirements.txt
pytest==8.3.5
requests==2.32.3
python-dotenv==1.0.1
openai>=1.0.0